Moved all integrations tests and files to src/integrationTest
