+++ build-common.gradle
@@ -31,27 +31,11 @@ ext.apacheCommonsNetVersion = '3.3'
 // PROJECT DIRECTORIES
 // ==============
 sourceSets {
-    test {
-        java {
-            exclude 'nl/esciencecenter/xenon/integration/**'
-            exclude '**/IT*.java'
-            exclude '**/*IT.java'
-            exclude 'nl/esciencecenter/xenon/adaptors/*/*AdaptorTest.java'
-        }
-    }
     integrationTest {
         java {
             compileClasspath += main.output + test.output
             runtimeClasspath += main.output + test.output
-            // TODO move all integration tests to src/integrationTest/java
-            srcDirs 'src/test/java'
-            include 'nl/esciencecenter/xenon/integration/**'
-            include '**/IT*.java'
-            include '**/*IT.java'
-            include 'nl/esciencecenter/xenon/adaptors/*/*AdaptorTest.java'
         }
-        // TODO move all integration resources to src/integrationTest/resources
-        resources.srcDirs 'src/test/resources'
     }
     userguide {
         java {
