+++ src/integrationTest/java/nl/esciencecenter/xenon/adaptors/gridengine/GridEngineJobAdaptorTest.java
@@ -23,10 +23,8 @@ import static org.junit.Assume.assumeTrue;
 import java.io.OutputStream;
 
 import nl.esciencecenter.xenon.adaptors.GenericJobAdaptorTestParent;
-import nl.esciencecenter.xenon.adaptors.scripting.SchedulerConnection;
 import nl.esciencecenter.xenon.files.OpenOption;
 import nl.esciencecenter.xenon.files.Path;
-import nl.esciencecenter.xenon.jobs.InvalidJobDescriptionException;
 import nl.esciencecenter.xenon.jobs.Job;
 import nl.esciencecenter.xenon.jobs.JobDescription;
 import nl.esciencecenter.xenon.jobs.JobStatus;
