+++ src/integrationTest/resources/scripts/create_symlinks
